Start new make3 build directory
The goal of make3 is to provide a build process that is friendly to
package maintainers for arbitrary Linux distributions. This is with the
hope that such a setup lowers the barrier for including J in the package
repositories of Linux distributions.

In particular, concrete goals of make3 are to correct the following
perceived deficiencies of make2:

1) Homegrown build system of shell scripts,
2) Fragile behaviour based on ad-hoc compiler detection,
3) Ignoring CFLAGS, LDFLAGS etc. of host system,
4) Non-error return on build failure, and
5) Unsavory reputation amongst package maintainers.

As an important final goal, we wish to *not break* builds on hosts that
currently work under the make and make2 workflows.