make3:libj,jconsole: Replace cd path hopping with explicit path variables
Hopping directories within a script both makes it easy to get lost and
is completely unnecessary if one uses appropriate command line options.
Particularly in this case, we have the -C flag to `make` that allows us
to set the working directory.

Additionally, the *ad hoc* realpath() function is superfluous given
`dirname` from coreutils, so we remove it.