make3:libj,jconsole: Set shell "strict mode"
This is a practice in defensive programming. We set the following three
shell options:

1) errexit   - Exit immediately if a command exits with non-zero status,
2) nounset   - Treat unset variables as an error when substituting, and
3) noclobber - Disallow output redirection from clobbering regular files.

In cases where compilation is part of a larger build chain---*e.g.* when
packaging for a Linux distribution---errors in the build can silently
result in a broken build. With `errexit` we immediately halt and let
upstream know of the problem.

The `nounset` option allows us to explicity differentiate between cases
we expect potentially-unset variables and cases we expect them to never
be empty. In conjunction with `errexit`, this lets functions as
automatic guards around variable use without the actual clutter of
explicit tests.

Similarly, the `noclobber` option forces us to be explicit in cases
where we *know* that file clobber is okay (see the >| redirect syntax).