make3:libj: Refactor logic into orthogonal units
The bulk of `build_libj.sh` is dedicated to setting up the environment
variables to be picked up by make. This commit performs a lot of clean
up regarding the logic of how said variables are set.

In particular, previously, much of the work was contained in
deeply-nested `if` blocks and one large `case` statement. By making use
of various forms of shell parameter expansion, we break up the
monolithic blocks of logic into units that for the most part work on one
variable at a time.

This, hopefully, makes the intent clearer. As a bonus, the
refactorization also decreases LOC by about half.