make3:libj,jconsole: Print all exported variables
In order to verify changes in the build system, I deterministically hash
(a tar of) the resulting outputs. The goal is to avoid introducing
unintended changes to the build.

In order to diagnose unintended hash-mismatches, however, we need to
compare the enviroments passed to make. This commit logs the relevant
environment variables.